+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.1.2
+
+- Add Global Broadcast capabilities:
+  - `globalBroadcast()`, to broadcast to all the globalConsume widgets.
+  - `globalConsumeAll(Widget Function() create, {Key? key})`, to create a widget which will be rebuild whenever any watched variables changes are made.
+  - `globalFrameAspects`, a getter, to return the current updating aspects.
+  - `globalAllAspects`, a getter, to return all the aspects that has been registered in the Global Mode.
+
 ## 2.1.1+3
 
 - Minor improvement.